refactor: mlflow utils
